fix: move mobile nav button next to note/file title
